J'ai essayé de répondre à ta demande de rétroaction de ma pull request.
