+++ API/models/DAOs/task.go
@@ -1,4 +1,4 @@
-package models
+package DAOs
 
 import (
 	"time"
