+++ API/repositories/task_repository.go
@@ -1,0 +1,11 @@
+package repositories
+
+import (
+	"llio-api/database"
+	"llio-api/models/DAOs"
+)
+
+// CreatTask insère une nouvelle tâche dans la BD
+func CreateTask(task *DAOs.Task) error {
+	return database.DB.Create(task).Error
+}
